Revamp of 'field' plugin.
Behind-the-scenes changes:
Instead of grabbing values ad-hoc, the field values
are grabbed at scan time and remembered in the %pagestate hash.
Calculated values are dealt with separately.
This requires all the back-ends to be changed!

The reasons for this change are:
a) more in keeping with IkiWiki workflow: get the data in the scan pass,
use it later, trusting it is all there.
b) Because data is saved in the %pagestate hash, they don't have to be
re-reckoned every time; this likely makes things faster.
c) makes it more possible to use other templating systems besides
HTML::Template in future; particularly HTML::Template::Pro, which
is faster, doesn't require a change of template files, but unfortunately
does not implement the "query" interface.

The disadvantage of this is that calculated values have to be treated
differently, and may not always work.