Dependencies, and state.
Trying to make this do the right thing in regard to dependencies;
it clears subset cache for pages which have been deleted or changed.
Also, I've moved the subset info back into the pagestate rather
than the wikistate, because if the subset page is deleted, the
state needs to go too.