Restructuring how field does its stuff.
Lots of changes to go back to the way it was designed originally;
to do "lazy evaluation" of field values when needed, rather than
calculating them beforehand.  This is done for two main reasons:
1. avoids calculations that don't need to be done
2. because field values are calculated when needed, it's much easier
to have field values that depend on other field values, rather
than having to worry about the order in which field scanning is done.

Note that field_get_value will return whatever type of value it
is given, rather than casting it to an array or scalar, so things
that use it should check what they're getting.

There have also been a number of tweaks to speed things up, after
profiling the code.