Added "newpage" plugin.
This adds a button + form to the actionbar, so that the user can
create a new page just by typing in the desired page name.
This is useful in the case where one doesn't want to create a link
to the new page from another page, because the new page will be
listed automatically anyway (such as with a map directive).
This is also more convenient than typing the page name in the
address bar and using the 404 plugin.