Add $escape to internal get_field_value call
get_other_page_field_value calls get_field_value, but the first parameter is now supposed to be an $escape string. The internal call was not including a value for this argument, so $field would get mapped to $escape and the remote field lookup fails.

This hack should be safe because we check the value of $escape at the beginning of the get_other_page_field_value (but I am no Perl programmer, so you should confirm this).